+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [3.5.1] - 2021-07-12
+### Changed
+- Show Patient id (dbGaP)
+- Store visible table columns for each service base URL separately
+
 ## [3.5.0] - 2021-07-07
 ### Added
 - Limit list of resources to what has data (dbGaP)